working on cleaning warnings

+++ src/Lyra2.c
@@ -168,7 +168,7 @@ int LYRA2(void *K, uint64_t kLen, const void *pwd, uint64_t pwdlen, const void *
     row = 0; //Resets the visitation to the first row of the memory matrix
     for (tau = 1; tau <= (int64_t) timeCost; tau++) {
     	//Step is approximately half the number of all rows of the memory matrix for an odd tau; otherwise, it is -1
-    	step = (tau % 2 == 0) ? -1 : nRows / 2 - 1;
+    	step = (tau % 2 == 0) ? -1 : (int64_t) nRows / 2 - 1;
     	do {
   	    //Selects a pseudorandom index row*
   	    //------------------------------------------------------------------------------------------
@@ -337,7 +337,7 @@ int LYRA2_old(void *K, uint64_t kLen, const void *pwd, uint64_t pwdlen, const vo
     row = 0; //Resets the visitation to the first row of the memory matrix
     for (tau = 1; tau <= (int64_t) timeCost; tau++) {
     	//Step is approximately half the number of all rows of the memory matrix for an odd tau; otherwise, it is -1
-    	step = (tau % 2 == 0) ? -1 : nRows / 2 - 1;
+    	step = (tau % 2 == 0) ? -1 : (int64_t) nRows / 2 - 1;
     	do {
   	    //Selects a pseudorandom index row*
   	    //------------------------------------------------------------------------------------------
